Rename TagEditor2 to TagEditor
Remove old angular tag-editor component and tests
