+++ src/sidebar/index.js
@@ -194,10 +194,13 @@ function startAngularApp(config) {
     )
     .component('streamContent', require('./components/stream-content'))
     .component('svgIcon', wrapReactComponent(require('./components/svg-icon')))
-    .component('tagList', wrapReactComponent(require('./components/tag-list')))
     .component(
       'tagEditor',
       wrapReactComponent(require('./components/tag-editor'))
+    )
+    .component(
+      'tagList',
+      wrapReactComponent(require('./components/tag-list'))
     )
     .component('threadList', require('./components/thread-list'))
     .component('topBar', wrapReactComponent(require('./components/top-bar')))
